+++ tests/selenium_tests/agora_unit_tests.py
@@ -29,45 +29,45 @@ class AgoraUI(unittest.TestCase):
 
   # Field Validations.
   def test_contact_form_validations(self):
-    ContactFormValidations("Firefox", headless=True, instance="http://localhost:8000/")
+    ContactFormValidations("Firefox", headless=True, instance="http://agora-backend/")
     assert True
 
   def test_provider_form_validations(self):
-    ProviderFormValidations("Firefox", headless=True, instance="http://localhost:8000/")
+    ProviderFormValidations("Firefox", headless=True, instance="http://agora-backend/")
     assert True
 
   def test_resource_form_validations(self):
-    ResourceFormValidations("Firefox", headless=True, instance="http://localhost:8000/")
+    ResourceFormValidations("Firefox", headless=True, instance="http://agora-backend/")
     assert True
 
 
   # ListView operations.
   def test_contacts_operations(self):
-    ContactsOperations("Firefox", headless=True, instance="http://localhost:8000/")
+    ContactsOperations("Firefox", headless=True, instance="http://agora-backend/")
     assert True
 
   def test_providers_operations(self):
-    ProvidersOperations("Firefox", headless=True, instance="http://localhost:8000/")
+    ProvidersOperations("Firefox", headless=True, instance="http://agora-backend/")
     assert True
 
   def test_resources_operations(self):
-    ResourcesOperations("Firefox", headless=True, instance="http://localhost:8000/")
+    ResourcesOperations("Firefox", headless=True, instance="http://agora-backend/")
     assert True
 
 
   # Create new records.
   def test_contact_create(self):
-    contact = ContactCreate("Firefox", headless=True, instance="http://localhost:8000/")
+    contact = ContactCreate("Firefox", headless=True, instance="http://agora-backend/")
     contact.create_new_contact()
     assert True
 
   def test_provider_create(self):
-    provider = CreateProvider("Firefox", headless=True, instance="http://localhost:8000/")
+    provider = CreateProvider("Firefox", headless=True, instance="http://agora-backend/")
     provider.create_new_provider(required_only=False)
     assert True
 
   def test_resource_create(self):
-    resource = ResourceCreate("Firefox", headless=True, instance="http://localhost:8000/")
+    resource = ResourceCreate("Firefox", headless=True, instance="http://agora-backend/")
     resource.create_new_resource(required_only=False)
     assert True