+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.2](https://github.com/grnet/agora-sp/compare/v1.4.1...v1.4.2) - 2021-11-04
+
+### Added
+- Allow portfolioadmin to edit the eosc_id.
+- Allow portfolioadmin to publish and update resources and providers to eosc portal.
+- Add CA_BUNDLE parameter in each external request django backend does.
+- Add eosc-state field as an editable form entry for resources and providers.
+
+### Changed
+-  Metrics: UI/UX changes updating the search bar and the provider/resources/user layout.
+
 ## [1.4.1](https://github.com/grnet/agora-sp/compare/v1.4.0...v1.4.1) - 2021-09-15
 
 ### Added